Dispatch tests in different classes
Use multiple inheritance to prevent unittest's discovery to run TestDGA
http://stackoverflow.com/questions/1323455/python-unit-test-with-base-and-sub-class#answer-1323554
